Incorporated new text that came in from data, added open faq pdf that is specific to rrb, renamed the 'water-use-measurement' component to keep up with the name changes
