+++ Source/Rio.Web/src/app/app-routing.module.ts
@@ -37,7 +37,7 @@ import { CreateUserCallbackComponent } from './pages/create-user-callback/create
 import { UserEditAccountsComponent } from './pages/user-edit-accounts/user-edit-accounts.component';
 import { AboutComponent } from './pages/about/about.component';
 import { GeneralFaqComponent } from './pages/general-faq/general-faq.component';
-import { WaterUseMeasurementComponent } from './pages/water-use-measurement/water-use-measurement.component';
+import { WaterUseMeasurementComponent } from './pages/measuring-water-use-with-openet/measuring-water-use-with-openet.component';
 import { DisclaimerComponent } from './pages/disclaimer/disclaimer.component';
 
 const routes: Routes = [
@@ -74,7 +74,7 @@ const routes: Routes = [
   { path: "help", component: HelpComponent },
   { path: "platform-overview", component: AboutComponent},
   { path: "frequently-asked-questions", component: GeneralFaqComponent},
-  { path: "measuring-water-use", component: WaterUseMeasurementComponent},
+  { path: "measuring-water-use-with-openet", component: WaterUseMeasurementComponent},
   { path: "glossary", component: GlossaryComponent },
   { path: "login-callback", component: LoginCallbackComponent },
   { path: "create-user-callback", component: CreateUserCallbackComponent },
