+++ Source/Rio.Web/src/app/app.module.ts
@@ -58,7 +58,7 @@ import { CreateUserCallbackComponent } from './pages/create-user-callback/create
 import { UserEditAccountsComponent } from './pages/user-edit-accounts/user-edit-accounts.component';
 import { AboutComponent } from './pages/about/about.component';
 import { GeneralFaqComponent } from './pages/general-faq/general-faq.component';
-import { WaterUseMeasurementComponent } from './pages/water-use-measurement/water-use-measurement.component';
+import { WaterUseMeasurementComponent } from './pages/measuring-water-use-with-openet/measuring-water-use-with-openet.component';
 import { DisclaimerComponent } from './pages/disclaimer/disclaimer.component';
 
 @NgModule({
