+++ Source/Rio.Web/src/app/pages/landowner-dashboard/landowner-dashboard.component.html
@@ -248,7 +248,7 @@
               <tr>
                 <th class="text-sm-left pb-0 mb-0">
                   Usage to Date
-                  <a routerLink="/measuring-water-use" target="_blank" style="cursor:pointer" placement="top" ngbTooltip="Learn how water usage is measured"><i
+                  <a routerLink="/measuring-water-use-with-openet" target="_blank" style="cursor:pointer" placement="top" ngbTooltip="Learn how water usage is measured"><i
                     class="fas fa-question-circle small"></i></a> 
                 </th>
                 <td class="text-sm-right pb-0 mb-0">
@@ -299,7 +299,7 @@
           <div class="card">
             <div class="card-header">
               {{waterYearToDisplay}} Allocation and Usage
-              <a routerLink="/measuring-water-use" target="_blank" style="cursor:pointer" placement="top" ngbTooltip="Learn how water usage is measured" onmouseenter="$(this).tooltip('show')"><i
+              <a routerLink="/measuring-water-use-with-openet" target="_blank" style="cursor:pointer" placement="top" ngbTooltip="Learn how water usage is measured" onmouseenter="$(this).tooltip('show')"><i
                 class="fas fa-question-circle small"></i>
               </a>
             </div>
@@ -328,7 +328,7 @@
             </label>
           </div>
           <span class="align-middle">{{waterYearToDisplay}} Water Usage by Parcel </span>
-          <a routerLink="/measuring-water-use" target="_blank" style="cursor:pointer; vertical-align: middle;" placement="top" ngbTooltip="Learn how water usage is measured"><i
+          <a routerLink="/measuring-water-use-with-openet" target="_blank" style="cursor:pointer; vertical-align: middle;" placement="top" ngbTooltip="Learn how water usage is measured"><i
             class="fas fa-question-circle small"></i></a> 
           <a class="btn btn-rio btn-sm mr-3 float-right" 
              routerLink="/parcel-override-et-data/{{activeAccount.AccountID}}/{{waterYearToDisplay}}"
