+++ Source/Rio.Web/src/app/pages/measuring-water-use-with-openet/measuring-water-use-with-openet.component.html
@@ -1,0 +1,102 @@
+<nav aria-label="breadcrumb">
+    <ol class="breadcrumb">
+        <li class="breadcrumb-item active" aria-current="page">Measuring Water Use with OpenET</li>
+    </ol>
+</nav>
+<div class="container">
+    <app-alert-display></app-alert-display>    
+    <div class="row mb-2">
+        <div class="col">
+            <h2 class="d-inline-block">Measuring Water Use with OpenET</h2>
+        </div>
+    </div>
+    <div class="row">
+        <div class="col">
+            <p><strong>Tracking water use and demand</strong></p>
+            <p>Water use calculations on this trading platform are provided by <a href="https://www.etdata.org/" target="_blank">OpenET</a>
+               - a  web-based platform being developed to provide easily accessible, credible estimates of consumptive water use 
+               at the field scale. OpenET relies on publicly available satellite and weather station data to estimate 
+               evapotranspiration (ET) - water that is transferred from the land to the atmosphere via evaporation from 
+               bare soil and transpiration through plant leaves. For irrigated agricultural fields, ET is necessary for 
+               plant growth and food production.
+               <br>
+               <br>
+               While several methods already exist for measuring ET at the field scale, they’re expensive, and since they’re
+               calculated differently by every consulting group, there isn’t a lot of broad trust in the data.
+               <u>Rosedale-Rio Bravo chose to partner with OpenET because it is bringing leading scientists and
+                  some of the most commonly used approaches for estimating ET together in one place, helping to 
+                  create a shared basis for decision-making.
+               </u>  
+               <br>
+               <br>
+               When it launches publicly in 2021, OpenET will significantly increase access to ET data for individual farmers 
+               and landowners, giving them the opportunity to see the same data that water managers and regulators are using 
+               by providing low-cost, near real-time ET estimates through a widely accessible website. The OpenET platform is 
+               currently in a beta version, and data are being provided here for user testing purposes only. However, with 
+               further work on model intercomparison and refinements, the objective of OpenET is to provide broadly trusted 
+               ET estimates that are well-documented, easily retrieved, and truly representative of the best available science.
+               <br>
+               <br>
+               Rosedale Rio Bravo utilizes data from OpenET within this trading platform to account for consumptive use 
+               on each assessor’s parcel. This information allows the district to accurately account for and report the
+               water consumptively used by each parcel, and gives landowners the ability to accurately track their water
+               budgets throughout the year.              
+            </p>
+        </div>
+    </div>
+    <div class="row">
+        <div class="col">
+            <p><strong>The OpenET Team</strong></p>
+            <p>The OpenET project is being led by NASA, Desert Research Institute (DRI) and Environmental Defense Fund with 
+               in-kind support from Google Earth Engine. The technical team includes approximately 30 researchers and practitioners 
+               from NASA, DRI, the U.S. Department of Agriculture, the U.S. Geological Survey, University of Nebraska, University of 
+               Idaho, University of Wisconsin, University of Maryland, California State University Monterey Bay, Google Earth Engine 
+               and the web development firm HabitatSeven, with expertise in satellite-based estimation of evapotranspiration, cloud 
+               computing, and user-driven website design.
+            </p>
+        </div>
+    </div>
+    <div class="row">
+        <div class="col">
+            <p><strong>Multiple ET Models</strong></p>
+            <p>OpenET is providing evapotranspiration data derived from several models that integrate satellite and meteorological data 
+               to generate estimates at the field scale. Because they are estimates, and because the models utilize different methodologies, 
+               they can produce results that diverge from field measurement. However, recent advances in remote sensing of ET have led to 
+               significant improvements in our ability to automate these methodologies, and many of these methods are already informing water 
+               use decisions across the West.                               
+            </p>
+        </div>
+    </div>
+    <div class="row">
+        <div class="col">
+            <p><strong>Creating an "ensemble" value</strong></p>
+            <p>The OpenET project is integrating many of these advances to support further improvements in water management by 
+               increasing access to remotely sensed ET data, and building confidence in the data by developing an ensemble value 
+               that is endorsed by a community of scientists. The models selected for inclusion in the initial ensemble are already 
+               informing water use decisions across the West, but the OpenET project will make it much easier to compare the models 
+               at scale, allowing identification of consistent differences between the various methods and for the scientific 
+               community to further refine and improve the estimates.
+               <br>
+               <br> 
+               The OpenET team is initiating one of the largest model intercomparison and accuracy assessments conducted to date to 
+               improve ET estimates. The results will inform the calculation of the single ensemble value from the full suite of ET 
+               models.
+               <br>
+               <br>
+               The data in the accounting platform represents the OpenET team’s preliminary estimate of an ensemble value based on the 
+               best available science. However, the science will significantly advance over the coming year as the intercomparison 
+               effort progresses and additional lessons are learned.
+               <br>
+               <br>
+               For additional questions about the accuracy and appropriate uses of this ET data, please contact Robyn Grimm, Environmental 
+               Defense Fund, at <a href="mailto:rgrimm@edf.org">rgrimm@edf.org</a>.
+                
+            </p>
+        </div>
+    </div>
+    <div class="row">
+        <div class="col mb-2">
+            <strong>Learn more about OpenET at <a href="https://www.etdata.org/" target="_blank">etdata.org</a> or <a href="assets/main/files/openet-faq-rrb.pdf" target="_blank">download the OpenET FAQ</a></strong>
+        </div>
+    </div>
+</div>
