+++ Source/Rio.Web/src/app/pages/measuring-water-use-with-openet/measuring-water-use-with-openet.component.ts
@@ -2,8 +2,8 @@ import { Component, OnInit } from '@angular/core';
 
 @Component({
   selector: 'rio-water-use-measurement',
-  templateUrl: './water-use-measurement.component.html',
-  styleUrls: ['./water-use-measurement.component.scss']
+  templateUrl: './measuring-water-use-with-openet.component.html',
+  styleUrls: ['./measuring-water-use-with-openet.component.scss']
 })
 export class WaterUseMeasurementComponent implements OnInit {
