+++ Source/Rio.Web/src/app/shared/components/header-nav/header-nav.component.html
@@ -51,8 +51,8 @@
                     <a routerLink="/frequently-asked-questions" class="dropdown-item">
                         Frequently Asked Questions
                     </a>
-                    <a routerLink="/measuring-water-use" class="dropdown-item">
-                        Measuring Water Use
+                    <a routerLink="/measuring-water-use-with-openet" class="dropdown-item">
+                        Measuring Water Use with OpenET
                     </a>
                     <a routerLink="/disclaimer/false" class="dropdown-item">
                         Disclaimer
